+++ src/main/java/com/demo/pteam/security/jwt/exception/DisabledAccountException.java
@@ -1,0 +1,11 @@
+package com.demo.pteam.security.jwt.exception;
+
+public class DisabledAccountException extends RuntimeException {
+    public DisabledAccountException(String message) {
+        super(message);
+    }
+
+    public DisabledAccountException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
